Profile is ready

+++ app/src/main/java/com/example/poseexercise/data/plan/Exercise.kt
@@ -2,12 +2,12 @@ package com.example.poseexercise.data.plan
 
 import androidx.annotation.DrawableRes
 
-data class Exercise (
+data class Exercise(
     val id: Long,
     val name: String,
     @DrawableRes
     val image: Int?,
-    val calorie: Int,
+    val calorie: Double,
     val level: String
 )
